+++ katana/payload.py
@@ -143,6 +143,7 @@ FIELD_MAPPINGS = {
     'status': 's',
     'swap': 's',
     'system': 's',
+    'tags': 't',
     'terminate': 't',
     'token': 't',
     'total': 't',
@@ -476,6 +477,25 @@ class CommandPayload(Payload):
             payload.set('command/arguments', args)
 
         return payload
+
+    @property
+    def request_id(self):
+        """
+        Get current request ID from command arguments.
+
+        The ID is available for request, response and action commands.
+
+        :rtype: str
+
+        """
+
+        # For request and response meta is an argument
+        rid = self.get('command/arguments/meta/id', '')
+        if not rid:
+            # For action payloads meta is part of the transport
+            rid = self.get('command/arguments/transport/meta/id', '')
+
+        return rid
 
 
 class CommandResultPayload(Payload):