+++ tests/api/test_request.py
@@ -162,3 +162,36 @@ def test_api_request_new_param():
     # Check error when a parameter has inconsisten type and value
     with pytest.raises(TypeError):
         request.new_param('foo', value=True, type=TYPE_INTEGER)
+
+
+def test_request_log(mocker, logs):
+    SchemaRegistry()
+
+    values = {
+        'attributes': {},
+        'component': object(),
+        'request_id': 'TEST',
+        'path': '/path/to/file.py',
+        'name': 'dummy',
+        'version': '1.0',
+        'framework_version': '1.0.0',
+        'client_address': '205.81.5.62:7681',
+        'gateway_protocol': urn.HTTP,
+        'gateway_addresses': ['12.34.56.78:1234', 'http://127.0.0.1:80'],
+        }
+    request = Request(**values)
+
+    log_message = u'Test log message'
+    # When debug is false no logging is done
+    assert not request.is_debug()
+    request.log(log_message)
+    out = logs.getvalue()
+    # There should be no ouput at all
+    assert len(out) == 0
+
+    # Create an instance with debug on
+    request = Request(debug=True, **values)
+    assert request.is_debug()
+    request.log(log_message)
+    out = logs.getvalue()
+    assert out.rstrip().split(' |')[0].endswith(log_message)